fix structure to next.js
